Update to 26.2
