+++ gradle.properties
@@ -6,15 +6,15 @@ org.gradle.configuration-cache=false
 
 # Fabric Properties
 	# check these on https://fabricmc.net/develop
-	minecraft_version=26.1
+	minecraft_version=26.2
 	loader_version=0.19.3
 	loom_version=1.17-SNAPSHOT
 
 # Mod Properties
-	mod_version = 1.8.2-26.1
+	mod_version = 1.8.2-26.2
 	maven_group = one.pouekdev
 	archives_base_name = coordinatelist
 
 # Dependencies
-	fabric_version=0.145.1+26.1
-	midnightlib_version=1.9.3+26.1-fabric
+	fabric_version=0.158.0+26.2
+	midnightlib_version=1.9.3+26.2-fabric
