+++ src/main/java/one/pouekdev/coordinatelist/CList.java
@@ -1,10 +1,12 @@
 package one.pouekdev.coordinatelist;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 import eu.midnightdust.lib.config.MidnightConfig;
 import net.fabricmc.api.ModInitializer;
 import net.fabricmc.fabric.api.client.command.v2.ClientCommandRegistrationCallback;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import net.fabricmc.fabric.api.client.rendering.v1.level.LevelRenderEvents;
 
 public class CList implements ModInitializer{
     public static final String MOD_ID = "coordinatelist";
@@ -14,5 +16,6 @@ public class CList implements ModInitializer{
     public void onInitialize(){
         MidnightConfig.init(MOD_ID, CListConfig.class);
         ClientCommandRegistrationCallback.EVENT.register(new CListCommand());
+        LevelRenderEvents.COLLECT_SUBMITS.register(new CListWaypointRenderer()::render);
     }
 }
