+++ src/main/java/one/pouekdev/coordinatelist/mixin/CListChatGrabber.java
@@ -36,7 +36,7 @@ public abstract class CListChatGrabber{
             int y = Math.round(Float.parseFloat(numbersList.get(1)));
             int z = Math.round(Float.parseFloat(numbersList.get(2)));
             Component clickableMessage = Component.translatable("chat.create.waypoint.message").withStyle(ChatFormatting.GREEN).withStyle(style -> style.withClickEvent(new ClickEvent.RunCommand("/clist " + x + " " + y + " " + z)));
-            CListVariables.delayedEvents.add(new CListDelayedEvent(0.1f, () -> CListVariables.minecraftClient.gui.getChat().addClientSystemMessage(clickableMessage)));
+            CListVariables.delayedEvents.add(new CListDelayedEvent(0.1f, () -> CListVariables.minecraftClient.gui.hud.getChat().addClientSystemMessage(clickableMessage)));
         }
     }
 }
