+++ src/main/resources/fabric.mod.json
@@ -30,7 +30,7 @@
   "depends": {
     "fabricloader": ">=0.19.3",
     "fabric-api": "*",
-    "minecraft": "~26.1",
+    "minecraft": "~26.2",
     "java": ">=25",
     "midnightlib": ">=1.9.3"
   },
